cgame: fix broken build (lcc doesn't know ssize_t)

+++ src/gamelogic/cgame/cg_particles.c
@@ -705,7 +705,7 @@ static void CG_ParseValueAndVariance( char *token, float *value, float *variance
 /*
 CG_ParseParticle helpers
 */
-static void CG_CopyLine( int *i, char *toks, ssize_t num, size_t size, char **text_p )
+static void CG_CopyLine( int *i, char *toks, int num, size_t size, char **text_p )
 {
 	char *token;
